pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "langchain-community>=0.3.21",
    "langchain>=0.3.23",
    "langchain-google-genai>=2.1.3",
    "langgraph>=0.3.31",
    "markdown>=3.8",
    "markupsafe>=3.0.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.40",
    "trafilatura>=2.0.0",
    "openai>=1.75.0",
]
